Added the proper new placement to the `mmap_t` `index` method to ensure correctness for object lifetimes. This has the side-effect that objects are default-initialized.
This poses a potential issue if you have an object in which the default empty construction clears the fields of the structure.

A potential solution to this issue is to specify a constructor overload that takes a `std::nullptr_t` that does no initialization of any structure members, allowing for the new in place to
call that, leaving everything in memory untouched.

To that end this commit also introduces the `has_nullable_ctor` template to allow SFINAE to check for objects with that exact constructor pattern and to call that when using the new in-place call.