pointer_utils: Implement managedTuple_t type
This structure operates as a RAII pointer, containing one of each
parameter type in the "releaser" function's signature. It can
automatically convert them to references or pointers, as needed.

The use case of this structure is for libraries such as libwebp or
libpng, where the release functions take one or multiple pointers to
structures created by the caller.