pointer_utils: Implement naughtyPtr_t and nicePtr_t
These two implement shared pointers to types with naughty (non-void
returning) and nice (void-returning) deleters.

The subclassing is necessary because, in libraries like Fontconfig,
either the deleter misbehaves when faced with a null pointer
(breaking the STL's contract of always executing the deleter on
reset()/dtor), or the deleter doesn't return void.