threadPool_t: Fix race condition in the tests
At present, the condition variable can be woken up by two events:

- any of the workers returns
- an OS interrupt

In both cases, there's a window in which the thread can cause the
wakeup, but the return insertion has not yet taken place
(usually because the caller is the first enqueued thread, but
it can also happen if the second is the caller, and the insertion from
the first hasn't been committed). This will
cause pool.queue(threads) to return false
because there's no actual return value available.

Co-authored-by: dragonmux <git@dragonmux.network>