ci: Use Meson to distill the coverage file
Meson supports distilling the gcov/gcda/gcno files into a
Codecov-compatible coverage.xml file with the `coverage-xml` target.
Using that allows us to skip the dance ourselves.

However, we must run the target manually-- unsure of what Meson 1.1.0 is
doing here, but it doesn't find the target if run through meson compile,
as the docs say.